Splitted base view model
